added test_metadata.py, test_config.py
